% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalkernel.R
\name{evalkernel}
\alias{evalkernel}
\title{Evaluate a kernel function}
\usage{
evalkernel(
  u,
  kernel = c("epanechnikov", "gaussian", "triweight", "uniform", "triangle", "cosinus",
    "quartic")
)
}
\arguments{
\item{u}{numeric array}

\item{kernel}{character string specifying name of kernel}
}
\value{
array with same dimensions as \code{u} containing kernel evaluated at \code{u}.
Note: \code{NA} values of \code{u} evaluate to 0.
}
\description{
Evaluate a kernel function
}
