% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdensity-package.R
\docType{package}
\name{logdensity-package}
\alias{logdensity-package}
\alias{_PACKAGE}
\title{logdensity: Estimate the (Logarithm) of the Density and its Derivatives}
\description{
This package contains functions that estimate the logarithm of an unknown density function from iid data.
    The estimation strategy is based on a local polynomial appoximation to the log-density. The estimates behave well
    near the boundary of the support of the density and can be guaranteed to be nonnegative. Details can be found in
    Pinske and Schurter (2020) "Estimates of derivatives of (log) densities and related objects."
}
\references{
Pinkse, J. and Schurter, K. (2020) "Estimates of derivatives of (log) densities and related objects."
}
\keyword{internal}
