% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdensityfit.R
\name{logdensity.fit}
\alias{logdensity.fit}
\title{Estimate the derivatives of the log-density and the log-density itself}
\usage{
logdensity.fit(data, x, h, g, dg, m, minx, maxx, logf, exact, ...)
}
\arguments{
\item{data}{Numeric vector of observations}

\item{x}{Points at which to estimate (if shorter than h, recycled to length
of h)}

\item{h}{Bandwidth (if shorter than x, recycled to the length of x)}

\item{g}{Function u, zl, and zr that must be equal to an S-length vector of
0's at zl = max((minx-x)/h,-1) and zr = min((maxx-x),1), where S is the order
of the local polynomial approximation to the log-density. Function must be
vectorized so that g(u, zl, zr) returns a matrix that is length(u) by S.}

\item{dg}{The derivative of g with respect to u. Must be vectorized and
return a matrix that is length(u) by S.}

\item{m}{Kernel function used to compute the density. Can be a function,
symbol, or character string that matches the name of a function or one of the
kernels allowed in evalkernel. If m == "epanechnikov" and the polynomial
order is 1, an exact solution is computed. Otherwise, the estimate of the 
log-density involves numerical integration.}

\item{minx}{Lower bound of support of x}

\item{maxx}{Upper bound of support of x}

\item{logf}{If TRUE (default), computes the density and log-density, in
addition to its derivatives.}

\item{exact}{logical indicating whether an exact solution should be used 
(if available) or numerical integration. Exact solution currently only
available with kernel and local linear approximation.}

\item{...}{Further arguments supplied to g and dg}

\item{S}{Degree of polynomial expansion of log-density to be used with
default g. If user supplies g and dg, this argument is ignored without
warning.}
}
\value{
A numeric vector containing the log-density (if logf == TRUE) and its derivatives
}
\description{
Estimate the derivatives of the log-density and the log-density itself
}
