% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdensity.R, R/logdensityfit.R
\name{logdensity}
\alias{logdensity}
\alias{logdensity.fit}
\title{A local polynomial log-density estimator}
\usage{
logdensity(
  data,
  x,
  h,
  g,
  dg,
  m = "epanechnikov",
  minx = -Inf,
  maxx = Inf,
  S = 1,
  logf = TRUE,
  mc.cores = 1L,
  ...
)

logdensity.fit(data, x, h, g, dg, m, minx, maxx, logf, exact, ...)
}
\arguments{
\item{data}{numeric vector of observations}

\item{x}{points at which to estimate (if shorter than h, recycled to length
of h)}

\item{h}{bandwidth (if shorter than x, recycled to the length of x)}

\item{g}{function of \code{u}, \code{zl}, and \code{zr} that must be equal to an \code{S}-length vector of
0's at \code{zl = max((minx-x)/h,-1)} and \code{zr = min((maxx-x),1)}, where \code{S} is the order
of the local polynomial approximation to the log-density. Function must be
vectorized so that \code{g(u, zl, zr)} returns a matrix that is \code{length(u)} by \code{S}.}

\item{dg}{function that evaluates derivative of \code{g} with respect to \code{u}. Must be
vectorized and return a matrix that is \code{length(u)} by \code{S}.}

\item{m}{kernel function used to compute the density. Can be a function,
symbol, or character string that matches the name of a function or one of the
kernels allowed in evalkernel. If \code{m == "epanechnikov"} and the polynomial
order is 1, an exact solution is computed. Otherwise, the estimate of the 
log-density involves numerical integration.}

\item{minx}{lower bound of support of \code{x}}

\item{maxx}{upper bound of support of \code{x}}

\item{S}{degree of polynomial expansion of log-density to be used with
default \code{g}. If user supplies \code{g} and \code{dg}, this argument is ignored without
warning.}

\item{logf}{logical indicating whether the log-density should be compute,
in addition to its derivative(s).}

\item{mc.cores}{integer number of cores to use with mcmapply. If equal to 1
(default), mapply will be used to loop over \code{x}, instead.}

\item{...}{Further arguments supplied to \code{g} and \code{dg}}

\item{exact}{logical indicating whether an exact solution should be used 
(if available) or numerical integration. Exact solution currently only
available with kernel and local linear approximation.}
}
\value{
an object of class \code{logdensity} which inherits from \code{matrix}.
    The \code{S+1} by \code{length(x)} matrix of estimated log-densities (\code{NA} unless \code{logf}
    is \code{TRUE}) and derivatives has the following additional attributes:
    \code{x}       vector of points at which the estimates were computed
    \code{n}       number of non-missing observations used in estimation
    \code{h}       bandwidth(s) used
    \code{call}    matched call

A numeric vector containing the log-density (if \code{logf == TRUE}) and its derivatives
}
\description{
\code{logdensity.fit} is intended to be called from within
\code{logdensity}, after performing basic argument verification. Use
caution when calling \code{logdensity.fit} directly.
}
\examples{
dat <- rchisq(n = 100, df = 2)
x <- seq(from = 0, to = 2, length.out = 20)
ld <- logdensity(data = dat, x = x, h = 0.5, m = "epanechnikov", minx = 0, S = 1, logf = TRUE)
print(ld)
plot(ld)

}
\references{
Pinkse, J. and Schurter, K. (2020) "Estimates of derivatives of (log) densities and related objects."
}
\seealso{
\code{\link[parallel]{mcmapply}}
}
