% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlogdensity.R
\name{plot.logdensity}
\alias{plot.logdensity}
\title{Print function for objects of class "logdensity"}
\usage{
\method{plot}{logdensity}(z, main = NULL, xlab = gettextf("n = \%g", z$n), type = "l", ...)
}
\arguments{
\item{z}{a logdensity object}

\item{main}{an overall title for the plots}

\item{xlab}{a title for the x axes}

\item{type}{what type of plot should be drawn}

\item{...}{further arguments passed to plot}
}
\description{
Print function for objects of class "logdensity"
}
