% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printlogdensity.R
\name{print.logdensity}
\alias{print.logdensity}
\title{Print function for objects of class "logdensity"}
\usage{
\method{print}{logdensity}(
  z,
  digits = max(3L, getOption("digits") - 3L),
  max = 6 * ncol(m),
  quote = FALSE,
  ...
)
}
\arguments{
\item{z}{a logdensity object}

\item{digits}{minimal number of significant digits}

\item{max}{approximate max number of entries to be printed}

\item{quote}{logical, indicating whether strings should be printed with surrounding quotes}

\item{...}{Further arguments passed to print()}
}
\value{
Invisibly returns z
}
\description{
Print function for objects of class "logdensity"
}
