% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdensity-package.R
\docType{package}
\name{logdensity-package}
\alias{logdensity-package}
\title{The logdensity package}
\description{
A package for estimating (log) densities and their derivatives
}
\details{
This package contains functions that estimate the logarithm of an unknown density function from iid data.
The estimation strategy is based on a local polynomial appoximation to the logarithm of the density. The
estimates behave well near the boundary of the support of the density and can be guaranteed to be nonnegative.
Details can be found in Pinske and Schurter (2020).
}
\references{
Pinkse, J. and Schurter, K. (2020) ''Estimates of derivatives of (log) densities and related objects.''
}
\author{
Joris Pinkse and Karl Schurter (maintainer, \email{kschurter@psu.edu})
}
