% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdensity.R
\name{logdensity}
\alias{logdensity}
\title{A local polynomial log-density estimator}
\usage{
logdensity(
  data,
  x,
  h,
  g,
  dg,
  m = "epanechnikov",
  minx = -Inf,
  maxx = Inf,
  S = 1,
  logf = TRUE,
  mc.cores = 1L,
  ...
)
}
\arguments{
\item{data}{Numeric vector of observations}

\item{x}{Points at which to estimate (if shorter than h, recycled to length
of h)}

\item{h}{Bandwidth (if shorter than x, recycled to the length of x)}

\item{g}{Function u, zl, and zr that must be equal to an S-length vector of
0's at zl = max((minx-x)/h,-1) and zr = min((maxx-x),1), where S is the order
of the local polynomial approximation to the log-density. Function must be
vectorized so that g(u, zl, zr) returns a matrix that is length(u) by S.}

\item{dg}{The derivative of g with respect to u. Must be vectorized and
return a matrix that is length(u) by S.}

\item{m}{Kernel function used to compute the density. Can be a function,
symbol, or character string that matches the name of a function or one of the
kernels allowed in evalkernel. If m == "epanechnikov" and the polynomial
order is 1, an exact solution is computed. Otherwise, the estimate of the 
log-density involves numerical integration.}

\item{minx}{Lower bound of support of x}

\item{maxx}{Upper bound of support of x}

\item{S}{Degree of polynomial expansion of log-density to be used with
default g. If user supplies g and dg, this argument is ignored without
warning.}

\item{logf}{If TRUE (default), computes the density and log-density, in
addition to its derivatives.}

\item{mc.cores}{Integer number of cores to use with mcmapply. If equal to 1
(default), mapply will be used to loop over x, instead.}

\item{...}{Further arguments supplied to g and dg}
}
\value{
an object of class "logdensity" which is a list containing
logdensity   matrix of estimated log-densities (if logf=TRUE) and derivatives
x            a vector of points at which the estimates were computed
n            the number of non-missing observations used in estimation
h            the bandwidth(s) used
call         the matched call
}
\description{
A local polynomial log-density estimator
}
\examples{
dat <- rchisq(n = 100, df = 2)
x <- seq(from = 0, to = 2, length.out = 20)
logdensity(data = dat, x = x, h = 0.5, m = "epanechnikov", minx = 0, S = 1, logf = TRUE)
print(logdensity)
plot(logdensity)

}
