% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlogdensity.R
\name{plot.logdensity}
\alias{plot.logdensity}
\title{Plot density and log-density estimates}
\usage{
\method{plot}{logdensity}(x, density = FALSE, type = "l", xlab, ylab, ...)
}
\arguments{
\item{x}{a logdensity object}

\item{density}{logical indicating whether to plot density or log-density}

\item{type}{what type of plot should be drawn}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{...}{further arguments passed to plot.default}
}
\description{
Plot density and log-density estimates
}
\seealso{
\code{\link[graphics]{plot.default}}, \code{\link[stats]{integrate}}
}
