% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printlogdensity.R
\name{print.logdensity}
\alias{print.logdensity}
\title{Print log-density estimates}
\usage{
\method{print}{logdensity}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  max = 100,
  quote = FALSE,
  ...
)
}
\arguments{
\item{x}{a logdensity object}

\item{digits}{minimal number of significant digits}

\item{max}{approximate max number of entries to be printed}

\item{quote}{logical, indicating whether strings should be printed with surrounding quotes}

\item{...}{further arguments passed to \code{print.default}}
}
\value{
invisibly returns x
}
\description{
Print log-density estimates
}
\seealso{
\code{\link{print.default}}
}
